requirements.txt
flask==2.3.0
paddlepaddle==2.5.0
paddleocr==2.7.0
pillow==10.0.0